MAINT: refactor occu_utils.py
