+++ smol/moca/kernel/bias.py
@@ -15,7 +15,7 @@ import numpy as np
 from smol.cofe.space.domain import get_species
 from smol.moca.composition import get_oxi_state
 from smol.moca.metadata import Metadata
-from smol.moca.utils.occu import get_dim_ids_table, occu_to_counts
+from smol.moca.occu_utils import get_dim_ids_table, occu_to_counts
 from smol.utils.class_utils import class_name_from_str, derived_class_factory
