--- smol/moca/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Mathematics and occupancy treatment utilities for moca."""
