+++ tests/test_moca/test_occu_utils.py
@@ -6,7 +6,7 @@ import numpy.testing as npt
 import pytest
 
 from smol.capp.generate.random import _gen_unconstrained_ordered_occu
-from smol.moca.utils.occu import (
+from smol.moca.occu_utils import (
     delta_counts_from_step,
     get_dim_ids_by_sublattice,
     get_dim_ids_table,
